Moved the schema reader classes to a separate package
